Add AppendSlice().
